refactor(harness): Phase-4 fixes — argv rejection, lib purity, dead fixtures, README accuracy [story-h4]
F5 (P1): usage-reader.ts main() filtered out every --prefixed token
before the positional-count check, so a valid path plus an unknown
flag exited 0 silently — contradicting the boundary-hygiene claim.
Fixed: any unrecognized -- token (outside --story's own id argument)
now routes to usage text + non-zero exit, in both dispatch modes.
Added 2 integration test cases (valid positional + unknown flag;
valid story id + unknown flag).

F7 (P3): lib/validate-path.ts imported node:fs/node:path, violating
harness/README.md's "pure logic in lib/, zero fs/process imports"
convention. Relocated to tool level (harness/metrics/validate-path.ts,
alongside the entrypoints); no behaviour change.

F1 (P1): deleted harness/metrics/fixtures/commit-log.txt and
fixtures/session-attribution.jsonl — referenced by no test; the
inline fixtures in the test files already provide the same diversity
coverage.

F9 (P3 soft): README's price-map paragraph described a date-based
staleness tolerance applyPrices doesn't implement (it only checks for
a missing model-rate entry). Tightened wording to match shipped
behaviour; no date check added.

F6: regenerated docs/metrics/loop.csv last, after all plan edits
(h4's own plan_loc row: 248 -> 264, reflecting this slice's edits).

F2: DoR checklist — ticked Phase 3 and Phase 4 complete (2026-07-02).

Plan's Suggestion log gets the Phase 4 code-reviewer findings block
appended verbatim (21 findings; F3/F4/F8/F10 acknowledged, no scope
change this slice).
