+++ docs/metrics/loop.csv
@@ -15,7 +15,7 @@ D,313,1789,1,0.17495807713806596,false
 h1,363,1231,1,0.29488220958570266,false
 h2,214,327,1,0.654434250764526,false
 h3,383,535,1,0.7158878504672898,true
-h4,248,n/a,0,n/a,false
+h4,264,n/a,0,n/a,false
 maint-01,276,395,1,0.6987341772151898,true
 maint-02,109,182,1,0.5989010989010989,true
 maint-03,173,408,1,0.42401960784313725,true
