--- harness/metrics/fixtures/session-attribution.jsonl
@@ -1,3 +1,0 @@
-{"parentUuid":null,"isSidechain":false,"type":"assistant","uuid":"fixture-attr-1","timestamp":"2026-07-02T09:30:00.000Z","cwd":"/Users/fixture/Projects/accounting","sessionId":"fixture-session-in-window","version":"2.1.197","message":{"model":"claude-sonnet-5","usage":{"input_tokens":1500,"output_tokens":300,"cache_creation_input_tokens":100,"cache_read_input_tokens":200}}}
-{"parentUuid":"fixture-attr-1","isSidechain":false,"type":"assistant","uuid":"fixture-attr-2","timestamp":"2026-07-02T10:00:00.000Z","cwd":"/Users/fixture/Projects/accounting","sessionId":"fixture-session-in-window","version":"2.1.197","message":{"model":"claude-sonnet-5","usage":{"input_tokens":800,"output_tokens":150,"cache_creation_input_tokens":0,"cache_read_input_tokens":50}}}
-{"parentUuid":null,"isSidechain":false,"type":"assistant","uuid":"fixture-attr-3","timestamp":"2026-05-01T09:00:00.000Z","cwd":"/Users/fixture/Projects/accounting","sessionId":"fixture-session-out-of-window","version":"2.1.197","message":{"model":"claude-fable-5","usage":{"input_tokens":900,"output_tokens":180,"cache_creation_input_tokens":0,"cache_read_input_tokens":0}}}
