+++ harness/metrics/validate-path.ts
